CORE-76: storage: move InMemoryStore to main, objectify 'internal'
