+++ storage/src/main/scala/coop/rchain/storage/IStore.scala
@@ -1,5 +1,7 @@
 package coop.rchain.storage
 
+import coop.rchain.storage.internal._
+
 /** The interface for the underlying store
   *
   * @tparam C a type representing a channel
