+++ storage/src/main/scala/coop/rchain/storage/util/package.scala
@@ -1,5 +1,8 @@
 package coop.rchain.storage
 
+import java.nio.file.attribute.BasicFileAttributes
+import java.nio.file.{FileVisitResult, Files, Path, SimpleFileVisitor}
+
 package object util {
 
   /**
